Address review comment: move to deployModal directory
